+++ packages/_private/angular-demo/src/demoform/demoform.component.ts
@@ -1,5 +1,5 @@
+import type { SynChangeEvent, SynRange } from '@synergy-design-system/components';
 import { serialize } from '@synergy-design-system/components';
-import type { SynChangeEvent } from '@synergy-design-system/components';
 import { Component, ElementRef, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 
@@ -7,9 +7,11 @@ const initialData = {
   code: '',
   comment: '',
   date: '',
+  donations: '2000 4000',
   email: '',
   files: undefined,
   gender: '',
+  happiness: '5',
   name: '',
   newsletterAngular: false,
   newsletterBeta: false,
@@ -32,6 +34,8 @@ export class DemoForm {
 
   @ViewChild('form') form!: ElementRef<HTMLFormElement>;
 
+  @ViewChild('donationRef') donationRef!: ElementRef<SynRange>;
+
   formData!: FormGroup;
 
   private _initFormData() {
@@ -42,6 +46,16 @@ export class DemoForm {
 
   constructor(private fb: FormBuilder) {
     this._initFormData();
+  }
+
+  ngAfterViewInit() {
+    const formatter = new Intl.NumberFormat('de-DE', {
+      currency: 'EUR',
+      maximumFractionDigits: 0,
+      style: 'currency',
+    });
+
+    this.donationRef.nativeElement!.tooltipFormatter = value => formatter.format(value);
   }
 
   reset() {