+++ packages/_private/e2e-demo-test/src/test.spec.ts
@@ -1,9 +1,16 @@
-import { Locator, expect, test } from '@playwright/test';
-import type SynInput from '../../../components/src/components/input/input.component';
-import type SynTextarea from '../../../components/src/components/textarea/textarea.component';
-import type SynCheckbox from '../../../components/src/components/checkbox/checkbox.component';
-import type SynSwitch from '../../../components/src/components/switch/switch.component';
-import type SynSelect from '../../../components/src/components/select/select.component';
+import {
+  Locator,
+  Page,
+  expect,
+  test,
+} from '@playwright/test';
+import type {
+  SynCheckbox,
+  SynInput,
+  SynSelect,
+  SynSwitch,
+  SynTextarea,
+} from '@synergy-design-system/components';
 import { type Framework, frameworks } from '../frameworks.config';
 import TestPage from './test.page';
 
@@ -31,7 +38,7 @@ async function fillField(locator: Locator, value: string) {
   await locator.blur();
 }
 
-async function fillForm(form: TestPage) {
+async function fillForm(form: TestPage, page?: Page) {
   await form.gender.getByText('Female').check();
   await fillField(form.name, 'Maxim');
   await fillField(form.email, 'max@musterman.de');
@@ -46,6 +53,29 @@ async function fillForm(form: TestPage) {
   await fillField(form.passwordRecovery, '1234');
   await form.topics.click();
   await form.angular.click();
+  await form.topics.evaluate((role: SynSelect) => {
+    // eslint-disable-next-line no-param-reassign
+    role.open = false;
+    role.blur();
+  });
+
+  // Drag the happiness handle to a 9 of 10 rating
+  await form.happiness.dragTo(form.happiness, {
+    targetPosition: { x: 1000, y: 50 },
+  });
+
+  // Drag the donations handle to a - - 100% rating
+  if (page) {
+    const firstKnob = await form.donations.locator('.thumb').first().evaluate((knob: HTMLDivElement) => `#donations #${knob.id}`);
+    const firstTick = '#donations syn-range-tick:first-of-type';
+
+    await page.dragAndDrop(firstKnob, firstTick);
+
+    const lastKnob = await form.donations.locator('.thumb').last().evaluate((knob: HTMLDivElement) => `#donations #${knob.id}`);
+    const lastTick = '#donations syn-range-tick:last-of-type';
+
+    await page.dragAndDrop(lastKnob, lastTick);
+  }
 }
 
 async function checkInitialState(form: TestPage) {
@@ -59,6 +89,8 @@ async function checkInitialState(form: TestPage) {
   expect(await getInputValue(form.passwordRecovery)).toBe('');
   expect(await getInputValue(form.topics)).toEqual([]);
   expect(await getInputValue(form.additionalInfo)).toBe('');
+  expect(await getInputValue(form.happiness)).toBe('5');
+  expect(await getInputValue(form.donations)).toBe('2000 4000');
   (await Promise.all(form.allNews.map((news) => getCheckedValue(news))))
     .forEach((val) => expect(val).toBeFalsy());
 
@@ -91,7 +123,7 @@ const createTestCaseFor = (framework: Framework) => {
       await form.goto(getURL(availablePages.form));
 
       // fill-out the form correctly
-      await fillForm(form);
+      await fillForm(form, page);
 
       // submit valid form
       await form.reset.click();
@@ -121,11 +153,7 @@ const createTestCaseFor = (framework: Framework) => {
       expect(await form.passwordRecovery.getAttribute('data-user-invalid')).toBeFalsy();
     });
 
-    test('Form submit', async ({ page, browserName }) => {
-      if (browserName === 'webkit' && process.env.PORT === '5175') {
-        return; // somehow this test is very flaky in the combination of webkit and port react
-      }
-
+    test('Form submit', async ({ page }) => {
       const form = new TestPage(page);
       await form.goto(getURL(availablePages.form));
 
@@ -144,12 +172,15 @@ const createTestCaseFor = (framework: Framework) => {
       await checkInitialState(form);
 
       // fill-out the form correctly
-      await fillForm(form);
+      await fillForm(form, page);
 
       // submit valid form
       await form.submit.click();
 
       expect(submitted).toBe(true);
+
+      // Not part of validation, but happiness should always be tested :)
+      expect(await getInputValue(form.happiness)).toBe('9');
 
       expect(await form.form.evaluate((f) => (f as HTMLFormElement).checkValidity())).toBe(true);
     });