+++ packages/angular/src/components/index.ts
@@ -35,6 +35,8 @@ export { SynProgressRingComponent } from './progress-ring.component';
 export { SynRadioButtonComponent } from './radio-button.component';
 export { SynRadioComponent } from './radio.component';
 export { SynRadioGroupComponent } from './radio-group.component';
+export { SynRangeComponent } from './range.component';
+export { SynRangeTickComponent } from './range-tick.component';
 export { SynSelectComponent } from './select.component';
 export { SynSideNavComponent } from './side-nav.component';
 export { SynSpinnerComponent } from './spinner.component';